requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pypdf
langchain
langchain-huggingface
langchain-community
langchain-chroma
groq
python-dotenv
sentence-transformers
chromadb
